requirements.txt
supabase
psycopg2-binary
fastapi
uvicorn[standard]
supabase
groq
python-dotenv
playwright >=1.45
streamlit >=1.32
pillow
